ocr_pdf/requirements.txt
python-dotenv==1.0.1
mistralai==1.2.3

